+++ src/tree.h
@@ -366,6 +366,7 @@ typedef enum tree_kind {
    T_MATCH_SELECT,
    T_PROT_DECL,
    T_DUMMY_DRIVER,
+   T_GUARD,
 
    T_LAST_TREE_KIND
 } tree_kind_t;