+++ test/test_parse.c
@@ -3210,7 +3210,7 @@ START_TEST(test_guarded)
    fail_unless(tree_has_guard(s0));
 
    tree_t gref = tree_guard(s0);
-   fail_unless(tree_kind(gref) == T_REF);
+   fail_unless(tree_kind(gref) == T_GUARD);
    fail_unless(tree_ref(gref) == g);
 
    tree_t s1 = tree_stmt(b2, 1);